feat(cli,shared): create/import/export/remove custom themes from the CLI
Moves the theme module (types, built-ins, parser, editor-theme) from
apps/desktop/src/shared/themes to @superset/shared/themes with re-export
shims so the CLI reuses the exact same validation and normalization the
Appearance UI applies. Adds settings theme import/export/remove: export a
built-in as a starter, edit the JSON, import, set.
