+++ packages/shared/package.json
@@ -32,6 +32,18 @@
 			"types": "./src/media-files.ts",
 			"default": "./src/media-files.ts"
 		},
+		"./themes": {
+			"types": "./src/themes/index.ts",
+			"default": "./src/themes/index.ts"
+		},
+		"./themes/types": {
+			"types": "./src/themes/types.ts",
+			"default": "./src/themes/types.ts"
+		},
+		"./themes/utils": {
+			"types": "./src/themes/utils.ts",
+			"default": "./src/themes/utils.ts"
+		},
 		"./v2-only-user": {
 			"types": "./src/v2-only-user.ts",
 			"default": "./src/v2-only-user.ts"
@@ -168,12 +180,14 @@
 	},
 	"dependencies": {
 		"@date-fns/tz": "1.4.1",
+		"culori": "4.0.2",
 		"friendly-words": "1.3.1",
 		"rrule": "2.8.1",
 		"semver": "7.7.4",
 		"zod": "4.4.3"
 	},
 	"devDependencies": {
+		"@types/culori": "4.0.1",
 		"@superset/typescript": "workspace:*",
 		"@types/friendly-words": "1.2.2",
 		"@types/semver": "7.7.1",
